Split ExternType out into a standalone crate
As mentioned in trait/README.md, this is useful for implementing
ExternType for types in a library crate without requiring all users of
the library to link in the full cxx along with its C++ symbols.
