+++ Cargo.toml
@@ -21,6 +21,7 @@ default = ["cxxbridge-flags/default"] # c++11
 "c++20" = ["cxxbridge-flags/c++20"]
 
 [dependencies]
+cxx-trait = { version = "=1.0.32", path = "trait" }
 cxxbridge-macro = { version = "=1.0.32", path = "macro" }
 link-cplusplus = "1.0"
 
@@ -36,7 +37,7 @@ rustversion = "1.0"
 trybuild = { version = "1.0.33", features = ["diff"] }
 
 [workspace]
-members = ["demo", "flags", "gen/build", "gen/cmd", "gen/lib", "macro", "tests/ffi"]
+members = ["demo", "flags", "gen/build", "gen/cmd", "gen/lib", "macro", "tests/ffi", "trait"]
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
