+++ tests/ui/wrong_type_id.stderr
@@ -1,13 +1,13 @@
 error[E0271]: type mismatch resolving `<StringPiece as ExternType>::Id == (f, o, l, l, y, (), B, y, t, e, R, a, n, g, e)`
-   --> $DIR/wrong_type_id.rs:11:9
-    |
-11  |         type ByteRange = crate::here::StringPiece;
-    |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ expected a tuple with 15 elements, found one with 17 elements
-    |
-   ::: $WORKSPACE/src/extern_type.rs
-    |
-    | pub fn verify_extern_type<T: ExternType<Id = Id>, Id>() {}
-    |                                         ------- required by this bound in `verify_extern_type`
-    |
-    = note: expected tuple `(f, o, l, l, y, (), B, y, t, e, R, a, n, g, e)`
-               found tuple `(f, o, l, l, y, (), S, t, r, i, n, g, P, i, e, c, e)`
+  --> $DIR/wrong_type_id.rs:11:9
+   |
+11 |         type ByteRange = crate::here::StringPiece;
+   |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ expected a tuple with 15 elements, found one with 17 elements
+   |
+  ::: $WORKSPACE/src/extern_type.rs
+   |
+   | pub fn verify_extern_type<T: ExternType<Id = Id>, Id>() {}
+   |                                         ------- required by this bound in `verify_extern_type`
+   |
+   = note: expected tuple `(f, o, l, l, y, (), B, y, t, e, R, a, n, g, e)`
+              found tuple `(f, o, l, l, y, (), S, t, r, i, n, g, P, i, e, c, e)`
